requirements.txt
Package            Version
------------------ -------
alembic            1.13.1
blinker            1.7.0
click              8.1.7
colorama           0.4.6
Flask              3.0.1
Flask-Cors         4.0.0
Flask-JWT-Extended 4.6.0
Flask-Migrate      4.0.5
Flask-SQLAlchemy   3.1.1
greenlet           3.0.3
itsdangerous       2.1.2
Jinja2             3.1.3
Mako               1.3.0
MarkupSafe         2.1.4
marshmallow        3.20.2
packaging          23.2
pip                22.0.4
psycopg2           2.9.9
PyJWT              2.8.0
python-dotenv      1.0.1
setuptools         58.1.0
SQLAlchemy         2.0.25
typing_extensions  4.9.0
Werkzeug           3.0.1
